requirements.txt
# Taj Chat - Requirements
# Ultimate AI Video Creation Platform

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Async HTTP
aiohttp>=3.9.0
httpx>=0.26.0

# AI Providers
openai>=1.10.0
anthropic>=0.18.0
together>=0.2.0
huggingface-hub>=0.20.0
transformers>=4.37.0
diffusers>=0.25.0

# Video Processing
moviepy>=1.0.3
opencv-python>=4.9.0
Pillow>=10.2.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.25
redis>=5.0.1
celery>=5.3.6

# Utilities
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.3
rich>=13.7.0
typer>=0.9.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Monitoring
prometheus-client>=0.19.0
sentry-sdk>=1.39.0
